requirements.txt
altgraph==0.17.2
aniso8601==7.0.0
anyio==3.5.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
asyncio==3.4.3
attrs==21.4.0
auto-py-to-exe==2.20.1
awsebcli==3.20.6
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
botocore==1.29.99
bottle==0.12.21
bottle-websocket==0.2.9
CacheControl==0.12.11
cachetools==4.2.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.3
contourpy==1.0.6
cryptography==36.0.1
cx-Oracle==8.3.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.0.4
djangorestframework==3.13.1
dnspython==2.2.1
EB==0.1.5
Eel==0.14.0
email-validator==1.1.3
entrypoints==0.4
executing==0.8.3
fastapi==0.75.0
fcm-django==1.0.12
filelock==3.8.0
firebase-admin==5.2.0
fonttools==4.38.0
future==0.18.2
fyers-apiv2==2.0.5
getmac==0.8.3
gevent==21.12.0
gevent-websocket==0.10.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.2
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.13.0
httpie==3.1.0
httplib2==0.20.4
httptools==0.2.0
idna==2.10
ipykernel==6.9.2
ipython==8.1.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.4
loguru==0.6.0
lxml==4.8.0
lz4==3.1.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mod-wsgi==4.9.3
msgpack==1.0.4
multidict==6.0.2
multitasking==0.0.11
nest-asyncio==1.5.4
netmiko==4.2.0
nsepy==0.8
ntc-templates==3.3.0
numpy==1.22.2
O365==2.0.19
oauthlib==3.2.0
opencv-python==4.6.0.66
orjson==3.6.7
outcome==1.1.0
packaging==21.3
pandas==1.4.1
paramiko==3.1.0
parso==0.8.3
pathspec==0.10.1
pbkdf2==1.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.4
promise==2.3
prompt-toolkit==3.0.28
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
py5paisa==0.5.3
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.0
Pygments==2.11.2
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyotp==2.8.0
pyOutlook==4.2.2
pyparsing==3.0.9
pypiwin32==223
PyQt6==6.2.3
PyQt6-Qt6==6.2.4
PyQt6-sip==13.2.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==22.3.0
razorpay==1.3.0
reportlab==3.6.8
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
Rx==1.6.1
scikit-learn==1.0.2
scipy==1.8.0
scp==0.14.5
selenium==4.1.2
semantic-version==2.8.5
singledispatch==3.7.0
six==1.14.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.34
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.17.1
stringcase==1.2.0
termcolor==1.1.0
text-unidecode==1.3
textfsm==1.1.3
threadpoolctl==3.1.0
tk==0.1.0
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
truedata-ws==4.2.4
twilio==7.8.2
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.15.0
virtualenv==20.17.0
virtualenvwrapper-win==1.2.7
watchgod==0.8.1
wcwidth==0.1.9
websocket-client==1.2.1
websockets==10.2
whichcraft==0.6.1
whitenoise==6.0.0
win32-setctime==1.1.0
wsproto==1.1.0
yfinance==0.1.87
zope.event==4.5.0
zope.interface==5.4.0
